Improved DropShadows, all object types can have a drop shadow now. For Objects with no fill (eg Groups) the drop shadow is calculated from the content.
git-svn-id: svn://scribus.net/trunk/Scribus@20111 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pdfwriter.cpp
@@ -758,7 +758,7 @@ namespace Pdf
     void Writer::write(const QByteArray& bytes)
     {
         outStream.writeRawData(bytes, bytes.size());
-    }
+	}
     
     void Writer::write(const ResourceDictionary& dict)
     {
@@ -836,7 +836,7 @@ namespace Pdf
     {
         assert( CurrentObj == 0);
         CurrentObj = id;
-        while (XRef.length() <= id)
+		while (static_cast<uint>(XRef.length()) <= id)
             XRef.append(0);
         XRef[id] = bytesWritten();
         write(toPdf(id));
